requirements.txt
azure-functions
azure-keyvault-secrets
azure-identity
azure-core
urllib3
responses
mock